requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
aiosqlite==0.19.0
alembic==1.13.1
shortuuid==1.0.11
